refactor(tarko): extract duration utils and remove code duplication
