+++ multimodal/tarko/agent-web-ui/src/common/utils/duration.ts
@@ -1,0 +1,14 @@
+/**
+ * Format duration from milliseconds to human readable string
+ */
+export const formatDuration = (ms: number): string => {
+  if (ms < 1000) {
+    return `${Math.round(ms)}ms`;
+  } else if (ms < 60000) {
+    return `${(ms / 1000).toFixed(1)}s`;
+  } else {
+    const minutes = Math.floor(ms / 60000);
+    const seconds = ((ms % 60000) / 1000).toFixed(1);
+    return `${minutes}m ${seconds}s`;
+  }
+};
